+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -5,7 +5,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 import java.util.List;
-import kitchenpos.application.support.MockMenuGroupDao;
+import kitchenpos.application.support.dao.MockMenuGroupDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
 import org.assertj.core.api.SoftAssertions;
